Se corregio la corrupcion de sectores de galaxia en la vizlaulicación de estrellals (en realidad era que los particle system del star visualizer se salian de los limites) y aun me estoy peleando con las fuenstes ademas se estan haciendo stres test para android con 110,250 estellas en el peor escenario

+++ Assets/Scripts/Generators/GalaxyGenerator.cs
@@ -462,8 +462,8 @@ public class GalaxyGenerator : MonoBehaviour
 		Vector3 sectorOrigin,
 		Vector3 sectorSize,
 		Vector2Int SectorPos,
-		int starCount,
-		GalaxyTypes galaxy, int sectorIndex = 0
+		long starCount,
+		GalaxyTypes galaxy, long sectorIndex = 0
 	)
 	{
 		var stars = new List<StarData>();
